feat(plasma-new-hope): refactor calendar utils
